Pass links through callbacks. Fix #56.
`linkify()` now accepts an iterable of callbacks as a second argument.
Those callbacks can alter the atributes of the link or return None to
either remove an existing link (the delinkify case) or prevent a new
link from being created.

The signature of the callbacks must be:

    def cb(attrs, new):

The first argument is a dict of attributes of the link, including the
inner text of the link (in the key '_text'). The second argument is a
boolean indicating whether the link is newly created or not.

These callbacks must either return the `attrs` dict, either as-is or
with entries added, removed, or modified, or None. If any of the
callbacks returns None, the link will be removed (or not created) and
replaced with its original text value.

Callbacks are run in the order they are given in the iterable.