Allow to set a function for the allowed_tags
Unfortunately I don't know how to make without the @staticmethod decorator.

But this adds the posibility to give a function to determine if a tag is
allowed or not. This allows blacklisting of tags instead of the current
whitelisting. In case your blacklist is much shorter then your whitelist,
especially when you want to allow new or custom tags.