Fix linkify edge case with ending )
It's common to put urls in parentheses. Sometimes the url_re will pick
up end parentheses that are part of the parenthetical, but shouldn't be
part of the url. This fix alleviates that somewhat by checking to see if
the url has a ( and if not, stripping any ) from the end. This assumes
that urls won't end in ) without also having a ( in them.

This is based on a fix from Istvan Albert.

Fixes #190