Prevent HTMLTokenizer from unescaping entities
This overrides the HTMLTokenizer's .consumeEntity() method such that it doesn't
convert character entities.

This also fixes some other escaping/unescaping oddities so that the output of
bleach.clean() is more correct in regards to intended behavior.

One thing this breaks is the idempotent property for bleach.clean()--it's no
longer idempotent. Since it escapes text more correctly now and that's not an
idempotent transform, this is no longer idempotent.

For example, bleach.clean() can't differentiate between a user talking about
code and saying this:

   I like my html wrapped in <b>!

and this:

   I like my html escaped like this &lt;b&gt;!

I'm not sure why we thought bleach.clean() could ever be correct and idempotent.
Seems like that was an error.