Fix regressions from Bleach 2.0 rewrite
The Bleach 2.0 rewrite involved changing Bleach clean() from making
changes during the tokenizing step to operating as an html5lib filter
that runs between the parsing and serializing steps. There were some
consequences of that rewrite namely that the html5lib parser got to
"fix" the HTML before Bleach could strip/escape disallowed tags. Because
of that, there were a variety of regressions from Bleach 1.x behavior
and Bleach 2.x behavior. For example:

1. Added end tags:

   bleach.clean('<sarcasm>') -> '&lt;sarcasm&gt;&lt;/sarcasm&gt;'

2. Parsed malformed tags as comments:

   bleach.clean('</sarcasm>') -> ''

Et cetera.

This change fixes that in a few ways. First, it overrides emitCurrentToken()
in the tokenizer to know about the allowed tags and whether they're going
to get stripped or escaped. It does the stripping at this point and changes
tags to be escaped into character data. That way the parser doesn't do any
"fixing".

Second, it adds some additional scrutiny for the
"expected-closing-tag-but-got-char" parser error case. That gets kicked
up when the tag is malformed. The parser will treat this as a comment
and then Bleach .clean() will drop it. Instead, this fixes that to
turn that into character data to be escaped later.

Fixes #271
Fixes #279
Fixes #280