Drop get_recent_tag_string()
tagOpenState() is always called for every < that denotes a StartTag,
EndTag, or ParseError and never any other time. We can (ab)use that
fact to reset the input stream history and then we don't have
to search backwards in the history for the < to find the original
tag string which is super because that was fragile and not very
performant.