Add recognized_tags to Linker arguments
recognized_tags lets someone specify a different set of recognized tags
to the Linker when they're using it by itself. One use of this is if
you're doing a clean pass, storing that, and then linkifying on demand.
The recognized_tags argument lets you re-use the allowed tags from the
clean step and then there isn't any additional oddities occuring in
the linkify step.

Fixes #409