Only preserve actual, valid, character entities
Previously, we accepted and preserved "ambiguous" character entities,
which they were classified as if they had some characters matching
an entity prefix and at most one non matching characters.

Keeping some nonexistant character entities while escaping the ampersand
in others seemed like a strange thing to do, so I thought that we might as
well check for if it exists as a valid entity.