Make the documentation reproducible.
Whilst working on the Reproducible Builds effort [0] we noticed
that python-bleach could not be built reproducibly.

This is because the documentation included a default arguments that
was (originally) generated from a "frozenset" type which are iterated
over at runtime in a nondeterministic order.

This commit therefore ensures a regular expression is constructed in
a deterministic manner, rendering the documentation reproducible.

This bug was originally filed in Debian as #934120 [1]

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/934120