Fix pip-tools problem and rework requirements (#636)
This fixes the problem with pip-tools < 6.5.0 and pip 22 by adding
pip-tools to the dev requirements, setting it to 6.5.0, and installing
that before checking the requirements.

This also moves requirements checking code to the scripts/run_tests.sh
script so it's easier to do in a local dev environment.

This also rebuilds the dev requirements.