Redo dev dependencies
Previously, we maintained dev dependencies in a .in file and then used
pip-compile to generate a requirements file and then used that for
building dev environments and CI test environments.

We hit dependencies that have different dependencies depending on Python
version and interpreter. That breaks with the compiled requirements
method.

This redoes dev requirements so that they're an extra specified in
setup.py. We don't need to track dependencies of dependencies or hash
dev dependencies. We can update dev dependencies as we need to. Having
stable pinned versions reduces unexpected errors because the universe
changed.

Developers will need to do:

pip install -e '.[dev]'

now to pick up dev dependencies.