Fix character entity handling (#450)
This fixes a bug where things in the form of a character entity (starts
with &, some letters, and ends with ;) that aren't character entities
get mangled. Now they are correctly handled.

This also fixes a bug where things in the form of a character entity
that aren't character entities should have had the & escaped, but that
wasn't happening.

After this, Bleach does the following:

* valid character entities that start with a & and end with a ; are left
  alone
* things that start with a & and end with a ; and aren't valid character
  entities are treated as text and the & is escaped
* things that start with a & but don't end with a ; are treated as text
  and the & is escaped

This is easy to reason about and idempotent and doesn't suffer from
ambiguous edge cases where Bleach can't know what the user intended.