Convert tags, skip_tags, recognized_tags to sets; fix doctests; f-strings
This converts the "tags" argument to BleachHTMLParser to be a set.

This converts the "skip_tags" and "recognized_tags" to linkify things to
be sets.

This updates the documentation fixing example code so that tags,
skip_tags, and recognized_tags are all sets.

This also converts some string interpolation from %s style to f-strings.